Im very sleepy
Fetching data on user when they first launch the application and storing it in a current user singleton. I retreive the amount they are allowed to carry torch and added the system force drop functionality. It took so long to figure out a nice way to display it but I finally settled on a customized toast that persists until timer runs out.