Server updates
New features:
Server can handle all commands in the command list except for u# and a? has not been tested.

Bugfixes:
Fixed SQL problems with t+
The return value for the traveled distance in u? is now clipped to a certain size
distance values are in kilometer

Todo:
There is still a problem with distance calculation
Verify distance calculation